Access Configuration externally instead of via subclass
